# app/demo/docs/swagger.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "demo",
    "version": "v0.1.0"
  },
...
  "host": "localhost:5188",
  "basePath": "/api/v1",
  "schemes": [
    "http",
    "https"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/user/get": {
      "post": {
        "summary": "获取用户",
        "operationId": "User_Get",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/userGetRsp"
            }
          },
          "401": {
            "description": "Returned when the user does not login.",
            "schema": {
              "$ref": "#/definitions/swaggerResponseBody"
            }
          },
          "403": {
            "description": "Returned when the user does not have permission to access the resource.",
            "schema": {
              "$ref": "#/definitions/swaggerResponseBody"
            }
          },
          "404": {
            "description": "Returned when the resource does not exist.",
            "schema": {
              "$ref": "#/definitions/swaggerResponseBody"
            }
          },
          "599": {
            "description": "Returned when the business error happen.",
            "schema": {
              "$ref": "#/definitions/swaggerResponseBody"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/userGetReq"
            }
          }
        ],
        "tags": [
          "用户"
        ]
      }
    },
    "/user/mget": {
      "post": {
        "summary": "获取多个用户",
        "operationId": "User_MGet",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/userMGetRsp"
            }
          },
          "401": {
            "description": "Returned when the user does not login.",
            "schema": {
              "$ref": "#/definitions/swaggerResponseBody"
            }
          },
          "403": {
            "description": "Returned when the user does not have permission to access the resource.",
            "schema": {
              "$ref": "#/definitions/swaggerResponseBody"
            }
          },
          "404": {
            "description": "Returned when the resource does not exist.",
            "schema": {
              "$ref": "#/definitions/swaggerResponseBody"
            }
          },
          "599": {
            "description": "Returned when the business error happen.",
            "schema": {
              "$ref": "#/definitions/swaggerResponseBody"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/userMGetReq"
            }
          }
        ],
        "tags": [
          "用户"
        ]
      }
    },
    "/user/search": {
      "post": {
        "summary": "搜索用户",
        "operationId": "User_Search",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/userSearchRsp"
            }
          },
          "401": {
            "description": "Returned when the user does not login.",
            "schema": {
              "$ref": "#/definitions/swaggerResponseBody"
            }
          },
          "403": {
            "description": "Returned when the user does not have permission to access the resource.",
            "schema": {
              "$ref": "#/definitions/swaggerResponseBody"
            }
          },
          "404": {
            "description": "Returned when the resource does not exist.",
            "schema": {
              "$ref": "#/definitions/swaggerResponseBody"
            }
          },
          "599": {
            "description": "Returned when the business error happen.",
            "schema": {
              "$ref": "#/definitions/swaggerResponseBody"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/userSearchReq"
            }
          }
        ],
        "tags": [
          "用户"
        ]
      }
    }
  },
  "definitions": {
...
    "protobufAny": {
      "type": "object",
      "properties": {
        "@type": {
          "type": "string"
        }
      },
      "additionalProperties": {}
    },
    "rpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    },
    "swaggerResponseBody": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "detail": {
          "type": "string"
        },
        "metadata": {
          "type": "object",
          "additionalProperties": {
            "type": "string"
          }
        }
      }
    },
    "userGetReq": {
      "type": "object",
      "properties": {
        "uid": {
          "type": "string",
          "format": "int64"
        }
      }
    },
    "userGetRsp": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "format": "int64"
        },
        "zid": {
          "type": "string"
        },
        "nickname": {
          "type": "string"
        },
        "avatar": {
          "type": "string"
        }
      }
    },
    "userMGetReq": {
      "type": "object",
      "properties": {
        "uids": {
          "type": "array",
          "items": {
            "type": "string",
            "format": "int64"
          }
        }
      }
    },
    "userMGetRsp": {
      "type": "object",
      "properties": {
        "list": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/userUserInfo"
          }
        }
      }
    },
    "userSearchReq": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        }
      }
    },
    "userSearchRsp": {
      "type": "object",
      "properties": {
        "list": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/userUserInfo"
          }
        }
      }
    },
    "userUserInfo": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "format": "int64"
        },
        "zid": {
          "type": "string"
        },
        "nickname": {
          "type": "string"
        },
        "avatar": {
          "type": "string"
        }
      }
    }
  },
  "securityDefinitions": {
    "ApiKeyAuth": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header"
    }
  },
  "security": [
    {
      "ApiKeyAuth": []
    }
  ],
...
}